pyproject.toml
[project]
name = "doc-intelligence-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb[cpu]>=1.3.7",
    "fastapi>=0.124.4",
    "groq>=0.37.1",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pydantic>=2.12.5",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",  
    "uvicorn[standard]>=0.38.0"
]
